works for polars, pandas, pyarrow - only one unit

+++ narwhals/_pandas_like/dataframe.py
@@ -370,8 +370,8 @@ class PandasLikeDataFrame:
         plx = self.__narwhals_namespace__()
         return self.filter(~plx.any_horizontal(plx.col(*subset).is_null()))
 
-    def estimated_size(self) -> int | float:
-        return self._native_frame.memory_usage(index=False, deep=True).sum()
+    def estimated_size(self) -> np.int64 | np.float64:
+        return float(self._native_frame.memory_usage(index=False, deep=True).sum())
 
     def with_row_index(self, name: str) -> Self:
         row_index = create_compliant_series(
